+++ utils/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/KafkaFactory.scala
@@ -1,5 +1,5 @@
 package pl.touk.nussknacker.engine.kafka
 
 object KafkaFactory {
-  final val TopicParamName = "topic"
+  final val TopicParamName = "Topic"
 }